tools/requirements.txt
black==23.7.0
click==8.1.7
colorama==0.4.6
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pyserial==3.5
typing_extensions==4.7.1
vgamepad==0.0.8
